change project architecture
