+++ app/src/main/java/com/example/myfinance/feature/presentation/screens/incomes/IncomesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.myfinance.screens
+package com.example.myfinance.feature.presentation.screens.incomes
 
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.padding
@@ -10,30 +10,28 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import com.example.myfinance.ui.common.FinappListItem
+import com.example.myfinance.ui.components.AppListItem
 import com.example.myfinance.R
-import com.example.myfinance.domain.Category
-import com.example.myfinance.ui.common.FinappFAB
-import com.example.myfinance.ui.common.FinappTopBar
+import com.example.myfinance.feature.domain.Category
+import com.example.myfinance.ui.components.AppFAB
+import com.example.myfinance.ui.components.AppTopBar
 
 @Composable
 fun IncomesScreen(incomes: List<Category>, onHistoryClicked: () -> Unit) {
     Scaffold (
         topBar = {
-            FinappTopBar(
+            AppTopBar(
                 title = "Доходы сегодня",
                 rightButtonIcon = R.drawable.history,
                 rightButtonDescription = "История",
                 rightButtonAction = onHistoryClicked
             ) },
-        floatingActionButton = { FinappFAB() },
+        floatingActionButton = { AppFAB() },
         contentWindowInsets = WindowInsets.statusBars
     ) { innerPadding ->
         LazyColumn(modifier = Modifier.padding(innerPadding)) {
             item {
-                FinappListItem(
+                AppListItem(
                     leftTitle = "Всего",
                     rightTitle = "600 000 ₽",
                     listBackground = MaterialTheme.colorScheme.secondary,
@@ -47,16 +45,4 @@ fun IncomesScreen(incomes: List<Category>, onHistoryClicked: () -> Unit) {
             }
         }
     }
-}
-
-@Composable
-fun IncomeItem(income: Category) {
-    FinappListItem(
-        leftTitle = income.category,
-        leftSubtitle = income.comment,
-        rightTitle = formatNumber(income.amount),
-        rightIcon = ImageVector.vectorResource(R.drawable.light_arrow),
-        clickable = true,
-        onClick = { /* Действие */}
-    )
 }
