+++ app/src/main/java/com/example/myfinance/feature/presentation/screens/settings/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.myfinance.screens
+package com.example.myfinance.feature.presentation.screens.settings
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -17,19 +17,15 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.myfinance.ui.common.FinappListItem
-import com.example.myfinance.R
-import com.example.myfinance.domain.Settings
-import com.example.myfinance.ui.common.FinappTopBar
+import com.example.myfinance.feature.domain.Settings
+import com.example.myfinance.ui.components.AppTopBar
 
 @Composable
 fun SettingsScreen(settings: List<Settings>){
     Scaffold (
-        topBar = { FinappTopBar(title = "Настройки") },
+        topBar = { AppTopBar(title = "Настройки") },
         contentWindowInsets = WindowInsets.statusBars
     ) { innerPadding ->
         LazyColumn(modifier = Modifier.padding(innerPadding)) {
@@ -60,15 +56,4 @@ fun SettingsScreen(settings: List<Settings>){
             }
         }
     }
-}
-
-@Composable
-fun SettingsItem(settings: Settings) {
-    FinappListItem(
-        leftTitle = settings.name,
-        rightIcon = ImageVector.vectorResource(R.drawable.dark_arrow),
-        listHeight = 56,
-        clickable = true,
-        onClick = { /* Действие */}
-    )
 }
