+++ app/src/main/java/com/example/myfinance/ui/components/AppListItem.kt
@@ -1,4 +1,4 @@
-package com.example.myfinance.ui.common
+package com.example.myfinance.ui.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -25,7 +25,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
 @Composable
-fun FinappListItem(
+fun AppListItem(
     leftTitle: String,
     leftSubtitle: String = "",
     rightTitle: String = "",
