+++ app/src/main/java/com/example/myfinance/ui/utils/formatNumber.kt
@@ -1,0 +1,12 @@
+package com.example.myfinance.ui.utils
+
+import java.text.DecimalFormat
+import java.text.DecimalFormatSymbols
+
+fun formatNumber(number: Int): String {
+    val symbols = DecimalFormatSymbols().apply {
+        groupingSeparator = ' '
+    }
+    val formatter = DecimalFormat("#,###", symbols)
+    return formatter.format(number) + " ₽"
+}
